Refactor tenancy forms
